Otimizando os extratos

+++ src/main/java/com/rinha/dto/ExtractDTO.java
@@ -1,6 +1,5 @@
 package com.rinha.dto;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -12,8 +11,8 @@ public record ExtractDTO(
         List<TransactionDTO> ultimas_transacoes
 ) {
     public record SaldoDto(
-            BigDecimal total,
-            BigDecimal limite,
+            int total,
+            int limite,
             LocalDateTime data_extrato
     ) {}
 }
